Refactored Halstead Integration and fixed casting bug
testing the actual visittoken process for the halstead - expression token integration and cast the expression result of hdifficulty to a double

+++ eclipse-cs/net.sf.eclipsecs.sample/src/net/sf/eclipsecs/sample/checks/HalsteadMetrics.java
@@ -68,7 +68,7 @@ public class HalsteadMetrics extends AbstractCheck{
 	}
 	
 	public void setHalsteadDifficulty() {
-		hDifficulty = ((expression_counter.getNumUniqueOperators()/2) * expression_counter.getNumOperands())
+		hDifficulty = (double)((expression_counter.getNumUniqueOperators()/2) * expression_counter.getNumOperands())
 		        /expression_counter.getNumUniqueOperands();
 	}
